requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
Pillow==10.1.0
aws-xray-sdk==2.11.0
boto3==1.28.0
requests==2.31.0
modal==1.0.0
pandas==2.1.0
datasets==3.2.0